+++ src/main/java/za/ac/cput/service/IVacancyService.java
@@ -2,6 +2,6 @@ package za.ac.cput.service;
 
 import za.ac.cput.model.Vacancy;
 
-public interface IVacancyService extends IService<Vacancy, Integer>{
-    void deleteById(Integer vacancyId);
+public interface IVacancyService extends IService<Vacancy, Long>{
+    void deleteById(Long vacancyId);
 }
